GUI/requirements.txt
numpy==1.23.4
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
pyscreenshot==3.1
subprocess.run==0.0.8
thread6==0.2.0
tk==0.1.0
tkvideoplayer==2.3